[Common] add reference comparator plots accessor functions
Add to the ReferenceUtils.h header file the functions that allow to access
the current, reference and ratio plots from the TCanvas created by the
ReferenceComparartorPlot class.

This will allow for example custom checkers to access the plots.
